sidebar and widgets refactor -- remove CurrentRound widget as it's ambiguous
