+++ extension/src/widgets/index.js
@@ -1,5 +1,6 @@
 export * from './Leaderboard';
-export * from './CurrentRound';
-export * from './LastRound';
+export * from './CurrentRoundOverview';
+export * from './NextRoundArticlePicker';
+export * from './LastRoundOverview';
 export * from './ArticlePicker';
 export * from './Rules';
